Allow long instead of int for addSample(). Everything else is already long internally.

+++ de.tudarmstadt.ukp.dkpro.core-asl/trunk/de.tudarmstadt.ukp.dkpro.core.api.frequency/src/main/java/de/tudarmstadt/ukp/dkpro/core/api/frequency/util/FrequencyDistribution.java
@@ -89,7 +89,7 @@ public class FrequencyDistribution<T> {
 	    return this.freqDist.keySet();
 	}
 	
-	public void addSample(T o, int number) {
+	public void addSample(T o, long number) {
 		this.n = this.n + number;
 		if (freqDist.containsKey(o)) {
 			freqDist.put(o, freqDist.get(o) + number);
